+++ src/types.ts
@@ -1,9 +1,13 @@
+/* eslint-disable no-unused-vars */
+
 import type {
   FieldDefinition,
   KeyedObject,
   ObjectSchemaType,
   Reference,
   SanityClient,
+  SanityDocument,
+  SanityDocumentLike,
 } from 'sanity'
 
 export type Language = {
@@ -15,6 +19,15 @@ export type SupportedLanguages =
   | Language[]
   | ((client: SanityClient) => Promise<Language[]>)
 
+export type PluginCallbackArgs = {
+  sourceDocument: SanityDocument
+  newDocument: SanityDocument
+  sourceLanguageId: string
+  destinationLanguageId: string
+  metaDocumentId: string
+  client: SanityClient
+}
+
 export type PluginConfig = {
   supportedLanguages: SupportedLanguages
   schemaTypes: string[]
@@ -23,6 +36,8 @@ export type PluginConfig = {
   bulkPublish?: boolean
   metadataFields?: FieldDefinition[]
   apiVersion?: string
+  allowCreateMetaDoc?: boolean
+  callback?: ((args: PluginCallbackArgs) => Promise<void>) | null
 }
 
 // Context version of config
@@ -43,7 +58,42 @@ export type Metadata = {
   translations: TranslationReference[]
 }
 
+export type MetadataDocument = SanityDocumentLike & {
+  schemaTypes: string[]
+  translations: TranslationReference[]
+}
+
 export type DocumentInternationalizationMenuProps = {
   schemaType: ObjectSchemaType
   documentId: string
+}
+
+// Extend Sanity schema definitions
+export interface DocumentInternationalizationSchemaOpts {
+  documentInternationalization?: {
+    /** Set to true to disable duplication of this field or type */
+    exclude?: boolean
+  }
+}
+
+declare module 'sanity' {
+  interface ArrayOptions extends DocumentInternationalizationSchemaOpts {}
+  interface BlockOptions extends DocumentInternationalizationSchemaOpts {}
+  interface BooleanOptions extends DocumentInternationalizationSchemaOpts {}
+  interface CrossDatasetReferenceOptions
+    extends DocumentInternationalizationSchemaOpts {}
+  interface DateOptions extends DocumentInternationalizationSchemaOpts {}
+  interface DatetimeOptions extends DocumentInternationalizationSchemaOpts {}
+  interface FileOptions extends DocumentInternationalizationSchemaOpts {}
+  interface GeopointOptions extends DocumentInternationalizationSchemaOpts {}
+  interface ImageOptions extends DocumentInternationalizationSchemaOpts {}
+  interface NumberOptions extends DocumentInternationalizationSchemaOpts {}
+  interface ObjectOptions extends DocumentInternationalizationSchemaOpts {}
+  interface ReferenceBaseOptions
+    extends DocumentInternationalizationSchemaOpts {}
+  interface SlugOptions extends DocumentInternationalizationSchemaOpts {}
+  interface StringOptions extends DocumentInternationalizationSchemaOpts {}
+  interface TextOptions extends DocumentInternationalizationSchemaOpts {}
+  interface UrlOptions extends DocumentInternationalizationSchemaOpts {}
+  interface EmailOptions extends DocumentInternationalizationSchemaOpts {}
 }